Split bot-workspace out of its router (720 -> 66 lines)
The last of the six core-object interacts to still be a workspace. The
whole conversation surface was inlined in the component whose only job is
choosing between browsing and chatting: the chat window, send controls,
runtime settings, prompt preview. All of it moved to bot-workspace.vue.

  <x-gallery v-if="!selected" />     <x-workspace v-else />

Every interact is now a router:

  bot        720 -> 66      character  923 -> 59
  scenario   531 -> 54      reward    1097 -> 61
  dream       57 (already)  facet      204 (no chat surface, by design)

SESSION STATE MOVED TO THE STORE, the same reason as the other three.
Interact mode was `!currentBot && sessionChats.length === 0` -- a
conversation under way keeps you here after the Bot is cleared -- and that
second half was a private `sessionChatIds` ref that had to move out with
the workspace, leaving the router blind to its own condition. chatStore
owns it under the 'bot' scope.

THE STATUS BANNER MOVED TOO, which is where this differs from
scenario-interact. There the banner reports on browse, configure and play
alike, so it belongs to the frame. Bot's only setters are a failed send and
the prompt-preview copy -- both require a Bot you are already working with,
so it has nothing to say in front of the gallery. The workspace is
multi-root as a result (banner and grid are siblings), so the router passes
no class and the grid keeps its own min-h-0 flex-1.

BOTH GRAPH-WALK CONTRACTS HELD WITH NO EDIT. verifyEntityArtManager and
verifyNarrativeKit followed bot-interact -> bot-workspace on their own and
stayed green at 7/7. Under the filename-keyed versions this split would
have reported two regressions, exactly as the previous three did.

Verified the move rather than assuming it: a line-bag diff of the original
against the router plus workspace shows nothing missing except the session
lines rewritten above -- the five other differences are prettier reflowing
wrapped text onto single lines, checked individually. One extraction bug
caught this way too: the first slice swallowed the router's own closing
</section>, which prettier rejected outright.

vue-tsc exit 0; entity-art, narrative-kit, route-gallery, layout,
capture-group and no-promise-in-store-state all pass.
