+++ utils/scripts/route-gallery-baseline.json
@@ -23,7 +23,7 @@
   "managerTotal": 0,
   "managers": {},
   "interactLines": {
-    "bot-interact": 721,
+    "bot-interact": 66,
     "character-interact": 59,
     "dream-interact": 57,
     "facet-interact": 204,
